Move module tests to test/integration
